split up schema module
